Small fix to weight factors in climate matrix
The calculation of the weights factors for precipitation in the climate matrix contained an exception for pixels that are ice-free in both GCM snapshots. However, this exception looked for reference ice thicknesses of 0.1 m rather than zero (a remnant from the old ANICE input files), preventing it from being triggered. Fixed this.