INCL. TIME-DEPENDENT CLIMATE FORCING FROM A FILE
With this commit we include the option to use climate forcing from a
file with time-dependent variables on a regular lon-lat grid. For new
option C%choice_forcing_method == 'SMB_direct' these are yearly SMB and
T2m, and the SMB is directly applied. For new option
C%choice_forcing_method == 'climate_direct' these are monthly T2m and
Precip, and the ITM SMB module is called to calculate SMB. Note that
currently no lapse rate or bias corrections are possible! Also, it
would be wise to set your SMB timestep smaller than the timestep in
your forcing record.