Small bugfix
Fixed a small error in PICO (erroneously used Tk and Sk instead of Tk(x,y) and Sk(x,y) in the final melt calculation).

Fixed a small error in calving (removed the ice-moves-only-out-of-completely-filled-calving-front-cells condition in the ice thickness update), so that the MISMIPplus shelf no loner forms a weird ridge at the ice front.