INCL. OCEAN INFRASTRUCTURE, INITIAL CHANGES
This commit facilitates the use WOA temperature/salinity as a forcing for the BMB routines.
When the already present config-parameter 'choice_ocean_temperature_model_config' is set to 'WOA',
a specified PD file will be read into 'T_ocean' and 'S_ocean'. These variables, which will probably
not cover the whole ice grid (probably absent under the ice shelves), are extended to do so.
This yields variables 'T_ocean_ext' and 'S_ocean_ext'.
The idea is to do this using the routine specified in Jourdain et al. (2018), but this is still
imcomplete:
1) The original WOA depth is used. A conservative 1D remap is needed to convert this.
2) The number of iterations of extending the ocean T/S coverage is now arbitrary. It should stop
automatically when the whole grid is filled.
3) Warm water should not protrude over sills in the bedrock topography. This is not yet implemented.
For the PD observations it is sufficient to copy 'T_ocean_ext' and 'S_ocean_ext', to the final
variables 'T_ocean_corr_ext' and 'S_ocen_corr_ext'. For climate snaphots PI, warm and cold, a bias
correction is included.