Added high-resolution ocean data extrapolation
The ocean data extrapolation routines now run on a higher resolution than the ice model itself. A high-resolution geometry file (same format as the initial/PD geometry files) is read to facilitate this. Ocean data is mapped first from the global lat/lon grid to the high-resolution regional grid. The extrapolation is performed there, and the results are saved to a NetCDF file. When the model is run, it automatically inspects all available pre-processed NetCDF files to see if any match the current model settings; if so, it just reads the data from that file (to save on precious computation time).

The default value for the high resolution is 5 km. Changing this to 2 km gives almost exactly the same answer. I haven't tried 1 km yet but I don't expect it will make any difference.