Melt models fix
Some bugfixes to the plume, PICO, and PICOP models. They now deal better with complex and/or changing ice geometry, in both the MISMIP+ and Antarctica experiments.

Also fixed a small bug in the time stepping code, making the model run a bit faster and more stable.

Replaced all the MISMIP+ and MISOMIP1 config files.