Added ISMIP6 output, fixed projection bug
Added the option to generate ISMIP6-compliant output files.

Fixed a bugs in the oblique stereographic projection code, which made the inverse projection give very wrong answers for the correct values of alpha_stereo. However, since the config values of alpha_stereo were found by "reverse engineering" using the same bugged code, this effect cancelled out, so that it still gave good lat/lon values.

The projection for Greenland has been changed to the newest ISMIP standard. Before, we used an oblique stereographic projection with a pole at lat = 72 N, lon = 40 W. Now, we use a polar stereographic projection (so with the pole at the actual North Pole), the central meridian at lon = 45 W, and a standard parallel (beta = 90 deg - alpha) of 20 degrees.

For North America, Eurasia, and Antarctica, we now use a standard parallel of 19 degrees (which for Antarctica is the ISMIP standard). It is not entirely clear how much of a change from before this represents, as due to the code bug, the old "alpha_stereo" was not actually the same as alpha, so that the old values of ~165 were pretty much meaningless.

The projection parameters for each model region are now configurable.