Adjust inversion to work with spin-up
Reference topography for inversion is now PD, as this is the target
topo for both the first (calibration) and third (deglaciation) phases
of the new spin-up protocol.

Added a modified version of UFEMISM's bed roughness inversion, which
adopts some of the features of the various inversions in IMAU-ICE. It
is only meant to provide an inversion style with which I am already
familiar.

Added a little model time and time step display to the standard
output of the simulation. Probably won't work on Snellius tho, but it
is cute and helpful for debugging/development/peace of mind.