Fix bug: wrong loop index in PD_obs method (clim/ocean)
For the constant present-day climate method in both the climate and
ocean modules, the main loop went over the range i2:i2, which you can
imagine is a recipy for kaput. Fixed this in both modules, and did a
quick scan of the model code to search for similar errors. Nothing
else found.