Set up spin-up phase 3/3: termination
All ready for third run of spin-up: a follow up the previous glacial
semi-cycle run with transient forcing and GIA between LGM and PD. In
this simulation, the inverted bed roughness map is used as initial
condition, and is inverted to provide a new map that helps the ice
sheet arrive at a PD configuration. Same with the basal melt rates
under ice shelves. As the forcing is relatively constant during the
Holocene, this should generate a spun-up PD state that will not
experience strong shocks once it is run into the future.