Bring back strict bed roughness inversion condition
For the spin-up deglaciation, the basal sliding inversion should not
start trying to reduce the volume of the ice sheet too early, but
wait until the change in climate does its job first. For this, a very
strict condition of only adjusting the basal parameter if that point
is experiencing no useful change of thickness is applied now.