Add custom bed roughness inversion interval
Added the option to set a time interval where the inversion of bed
roughness is active; outside of this interval the inversion is not
performed even if the config option is on. This is useful for a key
section of the spinup protocol, to avoid dividing it into too many
simulations. The default values for this interval are ridiculously
huge, just in case you, the innocent user, forget to set them up.