Add smoothing of prescribed bed roughness
As an option in the config files, the prescribed bed roughness
read from a file can now be smoothed upon reading, using a
Gaussian smoothing and a custom radius in terms of ice grid cells.

This solves an issue that arises when using a bed roughness created
by a run that has a different resolution from the one reading the
file, e.g. when using the calibrated bed roughness from a simulation
at 40km as initial guess for a simulation at 20km. Without this
smoothing, the simulation simply does not run.

Now that this issue is fixed, calibrations at higher resultions,
e.g. 10, 8 km, can use an initial (very educated) guess of the bed
roughness, which allows for a convergence within ~10 kyr instead of
the usual 30-50 kyr, and that's a lot! Cheers.