Add separate name for GCM ocean variables in config file
The config files can specify the name used for the ocean temperature
and salinity variables inside an input NetCDF file. The same name was
used for both the PD observational data set and the fields from the
GCMs, which is kinda not very flexible and would require extra work
when getting files with a different convention for names. Fixed this
by defining separate config variables for PD and GCM ocean data.