Greenland spin-up
All the config files and Matlab scripts needed to perform the different versions of the 4-phase Greenland spin-up. For the Holocene, two versions are available. The regular version applies a retreat mask so that the ice margin in 1950 matches the present day, and includes the anthropogenic CO2 spike. The "noPD" version doesn't have the retreat mask or the CO2 spike, so that it ends up in a sort of pre-Industrial steady state, which can be used as an initial state for paleo simulations.