Improved CO2 forcing from file with irregular time axis
Improved the way CO2 at model time is calculated when the text file containing the record has an irregular time axis.

CO2 is now also calculated as the average value over the preceding 30 years. For paleo simulations this won't make a difference, but for historical period / future runs it means the global warming spike is a bit smoother and a bit later.