Add first attempt at inverting cavling rates.
Based on the absolute ice velocities (vertically averaged) at
the front. If the front is beyond the limit radius (for the CalvingMIP
experiments), then the calving rate is greater than the velocity there,
and viceversa. Idea is that calving rates match the front veolities at
radius = 750 km, assuming that if front velocities are equal and
opposite sign to the calving rate, equilibrium should result.