[WIP] write inverted ocean T and S after spinup & initialise inverted T and S for projections
Regarding writing inverted ocean T and S after spinup. Add a subroutine in het netcdf_output module
to write the inverted ocean T and S to a nc file but an error occured when adding attributes to the
zeta variable in the file to be created.

This commit also contains the addition of config variables and a subroutine in the ocean module
to initialise the model with ocean T and S from a file. Also work in progress.