Add increased friction for thin-ice areas
Now there's an config option to set a threshold for
ice thickness, under which the minimum limit to beta
is increased from 0 to a configurable cap. This way,
grounded ice is guaranteed to have some kind of friction
which prevents super-fast flow over steep slopes when
the ice column gets very very thin.