Fix bug in extrapolation of ocean: master-only filename
The name of the output folder and file was only being passed
to the master process, which left the other processes with
no value for that variable. The new netcdf module actually
allows all processes to call its routines, but without the
value for that variable errors where triggered. Fixed that
by allowing all processes to access the outout file name.