Add transpose of lat/lon for grid read from xy NetCDF file
When a variable is read from a NetCDF file that contains a
xy grid, the associated grid is read as well, for which a
pair of lat/lon fields is derived from the x and y values.
At the end of the reading, the variable is transposed to
match IMAU-ICE "ji" convention, but the lat/lon fields are
left untransposed. This caused errors whenever the variable
was used together with the lat/lon fields. Fixed this by
transposing the lat/lon fields as well for all of the
read_field_from_xy_file_* mid-level routines.