Add options to basal slide: phi(0-1) and target relax field
Replaced the parameterised target relaxation phi_fric field
during the inversion for a copy of the inverted field at some
user-chosen time. Idea is that after a while the inversion will
not improve things but simply push inverted values to their
limits in search for that impossible perfect fit. Thus, if the
target is set to some good-enough point, then the inversion
will try to not deviate too much from that field.

Also added the option to replace the tangent of the phi_fric
till friction angle in the ZI sliding law by the variable itself,
i.e. phi_fric should have values between 0 and 1, mimicking
Tim's approach and his "C_Coulomb".

Replaced the hardcoded 0.96% limit to the pore water pressure
in the Martin2011 hydro parameterisation to a configurable value.
Intial idea is being able to set it to 0, i.e., no pore water /
hydrology at all. Of course now it can be used for tuning as well.