Add use of calving front effective thcikness for fluxes
Now the effective thickness is used for the computation
of fluxes at the calving front, instead of the standard
thickness. The problem with the latter is that the resulting
fluxes are smaller than what one would expect, since the
ice thickness is spread over the entire grid, yielding to
a very thin layer. If fluxes are needed for anything (e.g.
continuous calving rate estimations), then those will be
underestimated.

Also added the correction of fluxes at the calving front
to prevent outfluxes into the ocean before the cell is
"full", i.e., before the sub-grid calving front advances
to the border of the cell, according to the floating
margin fraction.

In order for this to wrok smoothly, the determination
of margin fractions needs to be done also during model
initialisation, so the first call to the ice dynamics
solver already has this information. Added this too.