Added if-ice-present condition to geometry relaxation
During the relaxation, the reference PD geometry is
updated to match the newly relaxed one. However, if
the ice sheet advanced beyond the PD margins, then
that leak would be also included in this update.
Fixed this be requiring that the PD thickness already
has ice there before updating it, otherwise it remains
at 0.