Add run_BMB in initialisation
BMB needs to be computed during initialisation to get the correct dHi/dt for the first time step of each run.
Also added run_BMB in the main model time loop before the velocities are computed such that the correct BMB pattern is taken to compute the new velocity field. Before this commit, the outdated BMB pattern was taken.