[WIP] The great shuffle is working
Seperate the computation of t_next from the action of setting the flags
to run the models. First compute t_next, then set time to time+t_next
and then set the flags to run each model to true or false. The shuffle
is working well now. However, the restart of the velocities is not correct
and the velocities written to the help_field and restart are very different
at the same step in time.