[WIP] Fix position of next timestep determinination
The time step of the next step should be determined before computing
the ice thickness for the next time step. That is now fixed. Concerning
the restart, all files are correctly read in for the restart run. After
initialising, the difference between the "no restart" run and the restart
run is 1.5 millimeter in ice thickness at 4 gridcells in the peninsula
and maximum 80 m/year in velocity at an ocean gridcell.