[WIP] Velocity restart update SSA and DIVA
Implemented the IF statement to skip the viscosity iteration in solve_DIVA when reading in velocities from the restart. Need to figure out if this is possible because it might need the viscosity field as well which is computed during the iteration.

Implemented some IF statements in order to be able to read and write velocity data from and to output when running SSA only. (Was only implemented for SIA/SSA and DIVA)