Fix race condition, Hb shuffle and Hi initialisation
Found race condition in derivatives_and_grids_module. Before mapping
from one grid to the other, all processors should be in sync.

Hb was computed and updated after the ice thickness and the mask but
should be done before. Also, dHb_dt should be computed at the end of
the initialisation.

During initialisation, Hs, Hi and Hb were recomputed based on Hb + dHb
in the adapt_initial_geometry_from_restart_file subroutine. When dHb is
zero, this caused and error in ice thickness of 10^-13. When dHb is nonzero,
this is incorrect following the new shuffle method. This whole subroutine
can be skipped and is commented now.