Fix restart for IMAU-ITM model
IMAU-ITM was reading firn depth and melt from the previous year
from the restart file. However, the restart file contained the newest
firn depth and melt, not the one from the previous year. To restart,
you actually need the one from the previous year. So I added a variable
to contains the firn depth and melt of the previous year for the restart
file. Note that dt_SMB_config should be set to 1 year to make this model
work well in the way it is written now. There is room for improvement on
this part.