Restart ELRA with dHb_dt instead of dHb
ELRA is now restarting with dHb_dt instead of dHb because the dHb
is dependent on the size of previous timestep, which is unknown in
a restart run. Now it is restarting from dHb_dt and then dHb is computed
from that.
Still need to clean up the code and create a do_restart if statement
when reading dHb_dt. Now it reads restarting with dHb_dt when ELRA is
turned on and when restart with velocities is set to true. But you always
want to restart with velocities so for the work in progress its okay.