Working version! Big cleanup
This versions works for projections, Franka's idealised setup and
the spinup. All comments are cleaned up and removed CALL syncs
Gaussian floodfill subroutine since that subroutine should only
be called by de master.

Users of IMAU-ICE should use particular timesteps for depending on
the application. For a perfect restart, set dt_min and dt_max to the
smallest other timestep that you choose for the modules. When using
the SMB model IMAU-ITM, use a time step of 1 year. All timesteps
should always be a binair number, for example 0.125, 0.25 or
any other integer.