[WIP] predictor/corrector works with SIA/SSA
dHb_dt and dHi_dt are now initialised in the new subroutine called
"initialise_changing_geometry_from_restart_file" so that dHs_dt is
computed from the correct geometry (before dHi_dt was zero at the
start of the restart run). Attention: I should implement that dHi_dt
is the second last timestep when pc is used and the last timestep
when direct is used.

When the  pc method is used, the second last timestep u_vav_cx and
v_vav_cy are safed to the restart file. When direcht method is used,
these variables are safed for the last time step only when DIVA is
used.

This commit contains many print statements because I still need to
test pc in combination with DIVA and I need to conduct a final test
for all combinations.