Working restart for all dynamics and timestepping variations
The restart works for direct and pc in combination with SIA, SSA,
SIA/SSA and DIVA. This commit also contains a big cleanup of print
statements.

The restart still only works when dt_min and dt_max are equal and set
to the smallest dt of the other modules. For example if BMB is 0.25 and
the other dt's are bigger, choose 0.25 for dt_min and dt_max.