Flow line extrapolation added
I have added my work on the flowline extrapolation, that was adapted from UFEMISM and originally created by Jorjo. This method can e.g., extrapolate basal friction based on the flow direction.

Currently, it is not called in the model but I am sure it will be of use to someone else.