Change name of configuration variables
Config variables to restart are now called "do_restart_config" and
"do_restart_from_previous_timestep_config". The first one should be
set to TRUE when one wants to restart in general. The second one
should be set to TRUE when you restart from a previous timestep.
For example, the previous run was from 2024 to 2034 and this run is
from 2034 to 2044. Set this config variable to FALSE if you restart
from a spinup for example.