Fix runladdie system call to make it compatible with snellius
The previous commit crashed on Snellius because (probably) xarray tried to access multiple processors while these were reserved for IMAU-ICE.

In order to solve this, a config variable was added 'BMB_laddie_system'.

When this is set to 'local_mac' it still uses line 2974 (BMB_module) to activate conda environment, run laddie, and deactivate conda environment. In this commit, the conda source is hardcoded in the model, still need to think of a cleaner solution.

If the config variable is set to 'slurm_HPC' it calls laddie using the system call in line 2977 of the BMB_module. Here, we call laddie using 'srun --ntasks=1 --cpus-per-task=1 --cpu-bind=verbose' which ensures only 1 processor is used for laddie. Make sure to request enough tasks when you submit a job on snellius and run IMAU-ICE on all cores but one - such that this core is available for laddie. For example, request 32 tasks and run IMAU-ICE using:

srun --ntasks=31 --cpus-per-task=1 IMAU_ICE_program config_IMAUICE