Calving flux correction
Flow from floating ice to open ocean is only allowed once the floating pixel is filled.
Old implementation resulted in accumulating ice at the calving front because ice was only allowed to flow out if the float_margin_frac > 0.99.
New implementation: Part of the flux at the calving front is used to fill the calving front cell itself, the remaining flux (if any) is allowed to flow into the next ocean cell. To account for 2D flow, the x and y fluxes are corrected based on the ratio of the fluxes.